+++ internal/assertions/json_test.go
@@ -17,6 +17,110 @@ func TestJSONEq(t *testing.T) {
 	}
 }
 
+func TestJSONMarshalUnmarshalAs(t *testing.T) {
+	t.Parallel()
+
+	for tc := range jsonMarshalCases() {
+		t.Run(tc.name, tc.test)
+	}
+}
+
+// =======================================
+// Test JSONMarshalAsT / JSONUnmarshalAsT
+// =======================================
+
+func jsonMarshalCases() iter.Seq[genericTestCase] {
+	type canMarshalJSON struct {
+		A string `json:"a"`
+		B int    `json:"b"`
+		C int    `json:"c,omitempty"`
+	}
+	type cantMarshalJSON struct {
+		A string `json:"a"`
+		B int    `json:"b"`
+		C func() `json:"c,omitempty"` // this fails when marshaling
+	}
+
+	const jazon = `{"a":"x","b":1}`
+	valueCanDo := canMarshalJSON{A: "x", B: 1}
+	valueCantDo := cantMarshalJSON{A: "x", B: 1, C: func() {}}
+
+	return slices.Values([]genericTestCase{
+		{"can JSON", testAllMarshalAs(valueCanDo, jazon, true)},
+		{"can't JSON/marshal-fails", testAllMarshalAs(valueCantDo, jazon, false)},
+		{"can JSON/different-values", testAllMarshalAs(valueCanDo, `{"a": 1,"b":"x"}`, false)},
+	})
+}
+
+//nolint:thelper // false positive: this is not a helper
+func testAllMarshalAs[Doc Text, Object any](value Object, jazon Doc, success bool) func(*testing.T) {
+	return func(t *testing.T) {
+		t.Run("with JSONMarshalAsT", testJSONMarshalAsT(value, jazon, success))
+
+		t.Run("with JSONUnmarshalAsT", testJSONUnmarshalAsT(value, jazon, success))
+	}
+}
+
+func testJSONMarshalAsT[Doc Text, Object any](value Object, jazon Doc, success bool) func(*testing.T) {
+	if success {
+		return func(t *testing.T) {
+			t.Run("should marshal JSON", func(t *testing.T) {
+				t.Parallel()
+
+				mock := new(mockT)
+				res := JSONMarshalAsT(mock, jazon, value)
+				if !res {
+					t.Errorf("expected struct to marshal correctly as JSON: %#v <=> %s.Got %s", value, jazon, mock.errorString())
+				}
+			})
+		}
+	}
+
+	return func(t *testing.T) {
+		t.Run("should not marshal JSON", func(t *testing.T) {
+			t.Parallel()
+
+			mock := new(mockT)
+			res := JSONMarshalAsT(mock, jazon, value)
+			if res {
+				t.Errorf("expected struct NOT to marshal correctly as JSON: %#v != %s", value, jazon)
+			}
+		})
+	}
+}
+
+func testJSONUnmarshalAsT[Doc Text, Object any](value Object, jazon Doc, success bool) func(*testing.T) {
+	if success {
+		return func(t *testing.T) {
+			t.Run("should unmarshal JSON", func(t *testing.T) {
+				t.Parallel()
+
+				mock := new(mockT)
+				res := JSONUnmarshalAsT(mock, value, jazon)
+				if !res {
+					t.Errorf("expected json string to unmarshal correctly from JSON: %#v <=> %s. Got %s", value, jazon, mock.errorString())
+				}
+			})
+		}
+	}
+
+	return func(t *testing.T) {
+		t.Run("should not unmarshal JSON", func(t *testing.T) {
+			t.Parallel()
+
+			mock := new(mockT)
+			res := JSONUnmarshalAsT(mock, value, jazon)
+			if res {
+				t.Errorf("expected json string NOT to unmarshal correctly from JSON: %#v != %s", value, jazon)
+			}
+		})
+	}
+}
+
+// =======================================
+// Test JSONEq variants
+// =======================================
+
 // test all JSONEq variants with the same input (possibly converted).
 func testAllJSONEq(expected, actual string, success bool) func(*testing.T) {
 	//nolint:thelper // linter false positive: this is not a helper