+++ require/require_assertions.go
@@ -1638,6 +1638,76 @@ func JSONEqT[EDoc, ADoc Text](t T, expected EDoc, actual ADoc, msgAndArgs ...any
 	t.FailNow()
 }
 
+// JSONMarshalAsT wraps [JSONEq] after [json.Marshal].
+//
+// The input JSON may be a string or []byte.
+//
+// It fails if the marshaling returns an error or if the expected JSON bytes differ semantically
+// from the expected ones.
+//
+// # Usage
+//
+//	actual := struct {
+//		A int `json:"a"`
+//	}{
+//		A: 10
+//	}
+//
+//	assertions.JSONUnmarshalAsT(t,expected, `{"a": 10}`)
+//
+// # Examples
+//
+//	success: []byte(`{"A": "a"}`), dummyStruct{A: "a"}
+//	failure: `[{"foo": "bar"}, {"hello": "world"}]`, 1
+//
+// Upon failure, the test [T] is marked as failed and stops execution.
+func JSONMarshalAsT[EDoc Text](t T, expected EDoc, object any, msgAndArgs ...any) {
+	if h, ok := t.(H); ok {
+		h.Helper()
+	}
+	if assertions.JSONMarshalAsT[EDoc](t, expected, object, msgAndArgs...) {
+		return
+	}
+
+	t.FailNow()
+}
+
+// JSONUnmarshalAsT wraps [Equal] after [json.Unmarshal].
+//
+// The input JSON may be a string or []byte.
+//
+// It fails if the unmarshaling returns an error or if the resulting object is not equal to the expected one.
+//
+// Be careful not to wrap the expected object into an "any" interface if this is not what you expected:
+// the unmarshaling would take this type to unmarshal as a map[string]any.
+//
+// # Usage
+//
+//	expected := struct {
+//		A int `json:"a"`
+//	}{
+//		A: 10
+//	}
+//
+//	assertions.JSONUnmarshalAsT(t,expected, `{"a": 10}`)
+//
+// # Examples
+//
+//	success: dummyStruct{A: "a"} , []byte(`{"A": "a"}`)
+//	failure: 1, `[{"foo": "bar"}, {"hello": "world"}]`
+//
+// Upon failure, the test [T] is marked as failed and stops execution.
+func JSONUnmarshalAsT[ADoc Text, Object any](t T, expected Object, jazon ADoc, msgAndArgs ...any) {
+	if h, ok := t.(H); ok {
+		h.Helper()
+	}
+	if assertions.JSONUnmarshalAsT[ADoc, Object](t, expected, jazon, msgAndArgs...) {
+		return
+	}
+
+	t.FailNow()
+}
+
 // Kind asserts that the [reflect.Kind] of a given object matches the expected [reflect.Kind].
 //
 // Kind reflects the concrete value stored in the object. The nil value (or interface with nil value)